requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
distlib==0.3.9
Django==5.1.5
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
mysql-connector==2.2.9
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.1
whitenoise==6.9.0
